+++ Vungle/VungleInterstitialCustomEvent.m
@@ -29,7 +29,7 @@
 
 #pragma mark - MPInterstitialCustomEvent Subclass Methods
 
-- (void)requestInterstitialWithCustomEventInfo:(NSDictionary *)info
+- (void)requestInterstitialWithCustomEventInfo:(NSDictionary *)info adMarkup:(NSString *)adMarkup
 {
     self.placementId = [info objectForKey:kVunglePlacementIdKey];
 
@@ -54,25 +54,27 @@
         NSMutableDictionary *options = [NSMutableDictionary dictionary];
         
         // VunglePlayAdOptionKeyUser
-        if ([[NSUserDefaults standardUserDefaults] objectForKey:kVungleUserId]) {
-            NSString *userID = [[NSUserDefaults standardUserDefaults] objectForKey:kVungleUserId];
+        NSString *userId = [self.localExtras objectForKey:kVungleUserId];
+        if (userId != nil) {
+            NSString *userID = userId;
             if (userID.length > 0) {
                 options[VunglePlayAdOptionKeyUser] = userID;
             }
         }
         
         // Ordinal
-        if ([[NSUserDefaults standardUserDefaults] objectForKey:kVungleOrdinal]) {
-            NSNumber *ordinalPlaceholder = [NSNumber numberWithLongLong:[[[NSUserDefaults standardUserDefaults] objectForKey:kVungleOrdinal] longLongValue]];
+        NSString *ordinal = [self.localExtras objectForKey:kVungleOrdinal];
+        if (ordinal != nil) {
+            NSNumber *ordinalPlaceholder = [NSNumber numberWithLongLong:[ordinal longLongValue]];
             NSUInteger ordinal = ordinalPlaceholder.unsignedIntegerValue;
             if (ordinal > 0) {
                 options[VunglePlayAdOptionKeyOrdinal] = @(ordinal);
             }
         }
         
-        // FlexVieAutoDismissSeconds
-        if ([[NSUserDefaults standardUserDefaults] objectForKey:kVungleFlexViewAutoDismissSeconds]) {
-            NSTimeInterval flexDismissTime = [[[NSUserDefaults standardUserDefaults] objectForKey:kVungleFlexViewAutoDismissSeconds] floatValue];
+        NSString *flexVieAutoDismissSeconds = [self.localExtras objectForKey:kVungleFlexViewAutoDismissSeconds];
+        if (flexVieAutoDismissSeconds != nil) {
+            NSTimeInterval flexDismissTime = [flexVieAutoDismissSeconds floatValue];
             if (flexDismissTime > 0) {
                 options[VunglePlayAdOptionKeyFlexViewAutoDismissSeconds] = @(flexDismissTime);
             }
@@ -107,9 +109,11 @@
 
 - (void)vungleAdWillAppear
 {
-
     MPLogAdEvent([MPLogEvent adWillAppearForAdapter:NSStringFromClass(self.class)], self.placementId);
     [self.delegate interstitialCustomEventWillAppear:self];
+}
+
+- (void)vungleAdDidAppear {
     MPLogAdEvent([MPLogEvent adShowSuccessForAdapter:NSStringFromClass(self.class)], self.placementId);
     MPLogAdEvent([MPLogEvent adDidAppearForAdapter:NSStringFromClass(self.class)], self.placementId);
     [self.delegate interstitialCustomEventDidAppear:self];
@@ -145,4 +149,7 @@
     [self.delegate interstitialCustomEvent:self didFailToLoadAdWithError:error];
 }
 
+- (NSString *)getPlacementID {
+    return self.placementId;
+}
 @end