+++ Vungle/VungleRewardedVideoCustomEvent.m
@@ -32,7 +32,7 @@
     [[VungleRouter sharedRouter] initializeSdkWithInfo:parameters];
 }
 
-- (void)requestRewardedVideoWithCustomEventInfo:(NSDictionary *)info
+- (void)requestRewardedVideoWithCustomEventInfo:(NSDictionary *)info adMarkup:(NSString *)adMarkup
 {
     self.placementId = [info objectForKey:kVunglePlacementIdKey];
 
@@ -80,6 +80,7 @@
     MPLogAdEvent([MPLogEvent adLoadSuccessForAdapter:NSStringFromClass(self.class)], self.placementId);
     [self.delegate rewardedVideoDidLoadAdForCustomEvent:self];
 }
+
 - (void)vungleAdWillAppear
 {
     MPLogAdEvent([MPLogEvent adWillAppearForAdapter:NSStringFromClass(self.class)], self.placementId);
@@ -88,6 +89,13 @@
     MPLogAdEvent([MPLogEvent adDidAppearForAdapter:NSStringFromClass(self.class)], self.placementId);
     [self.delegate rewardedVideoDidAppearForCustomEvent:self];
 }
+
+- (void)vungleAdDidAppear {
+    MPLogAdEvent([MPLogEvent adShowSuccessForAdapter:NSStringFromClass(self.class)], self.placementId);
+    MPLogAdEvent([MPLogEvent adDidAppearForAdapter:NSStringFromClass(self.class)], self.placementId);
+    [self.delegate rewardedVideoDidAppearForCustomEvent:self];
+}
+
 - (void)vungleAdWillDisappear
 {
     MPLogAdEvent([MPLogEvent adWillDisappearForAdapter:NSStringFromClass(self.class)], self.placementId);
@@ -111,7 +119,6 @@
     [self.delegate rewardedVideoShouldRewardUserForCustomEvent:self reward:[[MPRewardedVideoReward alloc] initWithCurrencyAmount:@(kMPRewardedVideoRewardCurrencyAmountUnspecified)]];
 }
 
-
 - (void)vungleAdDidFailToLoad:(NSError *)error
 {
     MPLogAdEvent([MPLogEvent adLoadFailedForAdapter:NSStringFromClass(self.class) error:error], self.placementId);
@@ -122,6 +129,10 @@
 {
     MPLogAdEvent([MPLogEvent adShowFailedForAdapter:NSStringFromClass(self.class) error:error], self.placementId);
     [self.delegate rewardedVideoDidFailToPlayForCustomEvent:self error:error];
+}
+
+- (NSString *)getPlacementID {
+    return self.placementId;
 }
 
 @end