+++ Vungle/CHANGELOG.md
@@ -1,4 +1,10 @@
 ## Changelog
+  * 6.4.5.0
+    * This version of adapters has been certified with Vungle 6.4.5 and is compatible with iOS 13.
+    * Stop implementing deprecated request API.
+    * Prioritize reading data from `localExtras` for rewarded video.
+    * Add support for the newly-introduced Vungle's Medium Rectangle format demand.
+
   * 6.3.2.6
     * Allow supported mediated networks and publishers to opt-in to process a user’s personal data based on legitimate interest basis. More details [here](https://developers.mopub.com/docs/publisher/gdpr-guide/#legitimate-interest-support).